requirements.txt
# Use this command to install all the dependencies
# pip install -r requirements.txt

# Core frameworks
langchain==1.0.0
langgraph==1.0.0
langchain-community
langchain-core

# Connectors
langchain-openai
langchain-anthropic

# Vector stores
faiss-cpu
chromadb
sentence-transformers

# Data science stack
numpy
pandas
matplotlib
scikit-learn
yfinance

# UI to send message into the model
streamlit

# Convery .ipynb to .py
nbconvert

# Memory management in langGraph
langgraph-checkpoint<3.0.0
langgraph-checkpoint-sqlite<3.0.0

# LLM model
google-generativeai
langchain-google-genai


# Utility for loading API keys from .env
# python-dotenv

# MCP Python SDK
mcp
fastmcp
python-dotenv